pyproject.toml
[tool.poetry]
name = "graphrag-legal"
version = "0.1.0"
description = "Hybrid GraphRAG system for legal and financial case documents"
authors = ["Rajveer <pegasuss167@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

pymupdf = "^1.24.0"
pdfplumber = "^0.11.0"
pytesseract = "^0.3.10"
Pillow = "^10.3.0"
gliner = "^0.2.0"
neo4j = "^5.20.0"
langchain-neo4j = "^0.2.0"
rdflib = "^7.0.0"
pyshacl = "^0.25.0"
sentence-transformers = "^3.0.0"
faiss-cpu = "^1.8.0"
pykeen = "^1.10.0"
openai = "^1.30.0"
rank-bm25 = "^0.2.2"
sqlalchemy = "^2.0.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.0"
numpy = "^2.0.0"
pandas = "^2.2.0"
fastapi = "^0.135.2"
uvicorn = "^0.42.0"
jinja2 = "^3.1.6"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"